+++ Entity/Model/Business/TeacherObservation.cs
@@ -5,8 +5,8 @@ namespace Entity.Model.Business
     public class TeacherObservation : ABaseEntity
     {
 
-        public long TeacherId { get; set; }
-        public long AgendaDayStudentId {  get; set; }
+        public int TeacherId { get; set; }
+        public int AgendaDayStudentId {  get; set; }
         public string Text { get; set; }
 
         public Teacher Teacher { get; set; }
